Transferred validate method to System class and added loadUsers method which loaded users stored in users.txt and stores them in users ArrayList as User objects.
